server/requirements.txt
python-dotenv
fastapi[standard]
sqlmodel
alembic
asyncpg
psycopg2
bcrypt
pyjwt